update Operations module
